style(Catalog): Rename service to repository
Service classes that we have for now contains only CRUD operations.
In order to have better maintainability I've changed names from 'Service' suffix to 'Repository' suffix in order to leave these names for class, that will contain more sophisticated business logic.
The following classes were renamed:
- ICategoriesService -> ICategoriesRepository
- CategoriesService -> CategoriesRepository
- IProductsService -> IProductsRepository
- ProductsSerivce -> ProductsRepository
All references to the renamed class were adjusted accordingly.
Affected classes:
- Catalog
  - Controllers
    - CategoriesController.cs
    - ProductsController.cs
  - Utils
    - ContainerBuilder.cs
- UnitTests\CatalogTests\ServiceTests
  - CategoriesRepositoryTests
  - ProductsRepositoryTests
